// tig-benchmarker/ui/src/app/interfaces/IImbalance.ts
export interface IImbalance {
  imbalance: number;
  c001_qualifiers: number;
  c002_qualifiers: number;
  c003_qualifiers: number;
  c004_qualifiers: number;
  c001_solutions: number;
  c002_solutions: number;
  c003_solutions: number;
  c004_solutions: number;
  deposit: number;
  deposit_qualifiers: number;
}